feat: Implement the foundational Golem AI agent system, introducing core components, multiple memory drivers, and conversation management.
