+++ src/core/Executor.js
@@ -1,0 +1,15 @@
+const { exec } = require('child_process');
+
+class Executor {
+    run(cmd) {
+        return new Promise((resolve, reject) => {
+            console.log(`⚡ Exec: ${cmd}`);
+            exec(cmd, { cwd: process.cwd() }, (err, stdout, stderr) => {
+                if (err) reject(stderr || err.message);
+                else resolve(stdout);
+            });
+        });
+    }
+}
+
+module.exports = Executor;
