+++ src/managers/SecurityManager.js
@@ -1,0 +1,22 @@
+// ============================================================
+// 🛡️ Security Manager (安全審計)
+// ============================================================
+// ==================== [KERNEL PROTECTED START] ====================
+class SecurityManager {
+    constructor() {
+        this.SAFE_COMMANDS = ['ls', 'dir', 'pwd', 'date', 'echo', 'cat', 'grep', 'find', 'whoami', 'tail', 'head', 'df', 'free', 'Get-ChildItem', 'Select-String', 'golem-check'];
+        this.BLOCK_PATTERNS = [/rm\s+-rf\s+\//, /rd\s+\/s\s+\/q\s+[c-zC-Z]:\\$/, />\s*\/dev\/sd/, /:(){:|:&};:/, /mkfs/, /Format-Volume/, /dd\s+if=/, /chmod\s+[-]x\s+/];
+    }
+    assess(cmd) {
+        const safeCmd = (cmd || "").trim();
+        const baseCmd = safeCmd.split(/\s+/)[0];
+        if (this.BLOCK_PATTERNS.some(regex => regex.test(safeCmd))) return { level: 'BLOCKED', reason: '毀滅性指令' };
+        if (this.SAFE_COMMANDS.includes(baseCmd)) return { level: 'SAFE' };
+        const dangerousOps = ['rm', 'mv', 'chmod', 'chown', 'sudo', 'su', 'reboot', 'shutdown', 'npm uninstall', 'Remove-Item', 'Stop-Computer'];
+        if (dangerousOps.includes(baseCmd)) return { level: 'DANGER', reason: '高風險操作' };
+        return { level: 'WARNING', reason: '需確認' };
+    }
+}
+// ==================== [KERNEL PROTECTED END] ====================
+
+module.exports = SecurityManager;
