+++ src/managers/ToolScanner.js
@@ -1,0 +1,20 @@
+const os = require('os');
+const { execSync } = require('child_process');
+
+// ============================================================
+// 🔍 ToolScanner (工具自動探測器)
+// ============================================================
+class ToolScanner {
+    static check(toolName) {
+        const isWin = os.platform() === 'win32';
+        const checkCmd = isWin ? `where ${toolName}` : `which ${toolName}`;
+        try {
+            const path = execSync(checkCmd, { encoding: 'utf-8', stdio: 'pipe' }).toString().trim().split('\n')[0];
+            return `✅ **已安裝**: \`${toolName}\`\n路徑: ${path}`;
+        } catch (e) {
+            return `❌ **未安裝**: \`${toolName}\`\n(系統找不到此指令)`;
+        }
+    }
+}
+
+module.exports = ToolScanner;
