+++ src/memory/ExperienceMemory.js
@@ -1,0 +1,33 @@
+const fs = require('fs');
+const path = require('path');
+
+// ============================================================
+// 🧠 Experience Memory (經驗記憶體)
+// ============================================================
+class ExperienceMemory {
+    constructor() {
+        this.memoryFile = path.join(process.cwd(), 'golem_learning.json');
+        this.data = this._load();
+    }
+    _load() {
+        try { if (fs.existsSync(this.memoryFile)) return JSON.parse(fs.readFileSync(this.memoryFile, 'utf-8')); } catch (e) { }
+        return { lastProposalType: null, rejectedCount: 0, avoidList: [], nextWakeup: 0 };
+    }
+    save() { fs.writeFileSync(this.memoryFile, JSON.stringify(this.data, null, 2)); }
+    recordProposal(type) { this.data.lastProposalType = type; this.save(); }
+    recordRejection() {
+        this.data.rejectedCount++;
+        if (this.data.lastProposalType) {
+            this.data.avoidList.push(this.data.lastProposalType);
+            if (this.data.avoidList.length > 3) this.data.avoidList.shift();
+        }
+        this.save();
+    }
+    recordSuccess() { this.data.rejectedCount = 0; this.data.avoidList = []; this.save(); }
+    getAdvice() {
+        if (this.data.avoidList.length > 0) return `⚠️ 注意：主人最近拒絕了：[${this.data.avoidList.join(', ')}]。請避開。`;
+        return "";
+    }
+}
+
+module.exports = ExperienceMemory;
